Integrate token unwrapping
