+++ packages/shared/src/lib/core/wallet/actions/index.ts
@@ -9,6 +9,5 @@ export * from './mintNativeToken'
 export * from './mintNft'
 export * from './rejectActivity'
 export * from './sendOutput'
-export * from './unwrapStardustAsset'
 
 export * from './send'
