+++ packages/shared/src/lib/core/wallet/actions/send/index.ts
@@ -1,3 +1,5 @@
+export * from './createEvmChainToEvmChainTransaction'
+export * from './createEvmChainToStardustNetworkTransaction'
 export * from './createEvmTransactionFromSendFlowParameters'
 export * from './createStardustOutputFromSendFlowParameters'
 export * from './sendOutputFromStardust'
