+++ packages/shared/src/lib/core/wallet/utils/send/index.ts
@@ -1,1 +1,2 @@
+export * from './buildTransferredAssetFromSendFlowParameters'
 export * from './validateSendConfirmation'
